move device to training
