+++ recipes/dev/generate_v2.py
@@ -81,7 +81,7 @@ class InferenceRecipe:
     """
 
     def __init__(self, cfg: DictConfig) -> None:
-        self._device = utils.get_device(device=cfg.device)
+        self._device = training.get_device(device=cfg.device)
         self._dtype = training.get_dtype(dtype=cfg.dtype, device=self._device)
         self._logger = utils.get_logger(cfg.log_level)
         training.set_seed(
@@ -127,7 +127,7 @@ class InferenceRecipe:
             f"Bandwidth achieved: {model_size * tokens_per_second / (1024**3):.02f} GiB/s"
         )
         if self._device.type != "cpu":
-            torch_device = utils.get_torch_device_namespace()
+            torch_device = training.get_torch_device_namespace()
             self._logger.info(
                 f"Max memory allocated: {torch_device.max_memory_allocated() / (1024**3):.02f} GiB"
             )
@@ -190,7 +190,7 @@ class InferenceRecipe:
             ).unsqueeze(0)
         batch["mask"] = causal_mask[None, :seq_len]
         batch["input_pos"] = input_pos[None, :seq_len]
-        utils.batch_to_device(batch, self._device)
+        training.batch_to_device(batch, self._device)
 
         # 6. Prefill step
         generated_tokens = []
