+++ recipes/knowledge_distillation_distributed.py
@@ -104,7 +104,7 @@ class KDRecipeDistributed(FTRecipeInterface):
     """
 
     def __init__(self, cfg: DictConfig) -> None:
-        self._device = utils.get_device(device=cfg.device)
+        self._device = training.get_device(device=cfg.device)
         # Reduced precision logic
         self._dtype = training.get_dtype(cfg.dtype, device=self._device)
         # fp16 precision is explicitly disabled as it is not supported in this
@@ -124,7 +124,7 @@ class KDRecipeDistributed(FTRecipeInterface):
         )
         init_process_group(self.distributed_backend)
 
-        self.world_size, self.rank = utils.get_world_size_and_rank()
+        self.world_size, self.rank = training.get_world_size_and_rank()
 
         self._is_rank_zero = self.rank == 0
 
@@ -818,7 +818,7 @@ class KDRecipeDistributed(FTRecipeInterface):
                 ):
                     torch.cuda.memory._record_memory_history()
 
-                utils.batch_to_device(batch, self._device)
+                training.batch_to_device(batch, self._device)
 
                 # Calculate the number of unmasked tokens in the current batch
                 # and increment the total number of tokens seen in the step
