+++ recipes/ppo_full_finetune_single_device.py
@@ -111,7 +111,7 @@ class PPOFullFinetuneRecipeSingleDevice(FTRecipeInterface):
 
     def __init__(self, cfg: DictConfig) -> None:
 
-        self._device = utils.get_device(device=cfg.device)
+        self._device = training.get_device(device=cfg.device)
         self._dtype = training.get_dtype(cfg.dtype, device=self._device)
 
         # Disable for fp16, as we haven't validated "full" fp16 with this recipe, nor
