+++ recipes/qat_distributed.py
@@ -131,7 +131,7 @@ class QATRecipeDistributed(FTRecipeInterface):
     """
 
     def __init__(self, cfg: DictConfig) -> None:
-        self._device = utils.get_device(device=cfg.device)
+        self._device = training.get_device(device=cfg.device)
         self._dtype = training.get_dtype(cfg.dtype, device=self._device)
 
         if self._dtype == torch.float16:
@@ -170,7 +170,7 @@ class QATRecipeDistributed(FTRecipeInterface):
             )
             self._log_peak_memory_stats = False
 
-        self.world_size, self.rank = utils.get_world_size_and_rank()
+        self.world_size, self.rank = training.get_world_size_and_rank()
         self._is_rank_zero = self.rank == 0
 
         # Training cfg
@@ -777,7 +777,7 @@ class QATRecipeDistributed(FTRecipeInterface):
                         )
                         self._model.apply(enable_fq)
 
-                utils.batch_to_device(batch, self._device)
+                training.batch_to_device(batch, self._device)
 
                 # Calculate the number of unmasked tokens in the current batch
                 # and increment the total number of tokens seen in the step
