+++ torchtune/training/_profiler.py
@@ -19,7 +19,8 @@ from omegaconf import DictConfig
 from torch._C._profiler import _ExperimentalConfig
 from torch.profiler import tensorboard_trace_handler
 
-from torchtune.utils import get_logger, get_world_size_and_rank
+from torchtune.training import get_world_size_and_rank
+from torchtune.utils import get_logger
 
 log = get_logger("INFO")
